+++ pyannote/metrics/segmentation.py
@@ -28,14 +28,17 @@
 # Camille Guinaudeau - https://sites.google.com/site/cguinaudeau/
 # Mamadou Doumbia
 # Diego Fustes diego.fustes at toptal.com
+from typing import Tuple, Union, Optional
 
 import numpy as np
 from pyannote.core import Segment, Timeline, Annotation
 from pyannote.core.utils.generators import pairwise
 
 from .base import BaseMetric, f_measure
+from .types import MetricComponents, Details
 from .utils import UEMSupportMixin
 
+#  TODO: can't we put these as class attributes?
 PURITY_NAME = 'segmentation purity'
 COVERAGE_NAME = 'segmentation coverage'
 PURITY_COVERAGE_NAME = 'segmentation F[purity|coverage]'
@@ -65,11 +68,13 @@ class SegmentationCoverage(BaseMetric):
 
     """
 
-    def __init__(self, tolerance=0.500, **kwargs):
-        super(SegmentationCoverage, self).__init__(**kwargs)
+    def __init__(self, tolerance: float = 0.500, **kwargs):
+        super().__init__(**kwargs)
         self.tolerance = tolerance
 
-    def _partition(self, timeline, coverage):
+    def _partition(self,
+                   timeline: Timeline,
+                   coverage: Timeline) -> Annotation:
 
         # boundaries (as set of timestamps)
         boundaries = set([])
@@ -85,13 +90,15 @@ class SegmentationCoverage(BaseMetric):
 
         return partition.crop(coverage, mode='intersection').relabel_tracks()
 
-    def _preprocess(self, reference, hypothesis):
+    def _preprocess(self, reference: Annotation,
+                    hypothesis: Union[Annotation, Timeline]) \
+            -> Tuple[Annotation, Annotation]:
 
         if not isinstance(reference, Annotation):
             raise TypeError('reference must be an instance of `Annotation`')
 
         if isinstance(hypothesis, Annotation):
-            hypothesis = hypothesis.get_timeline()
+            hypothesis: Timeline = hypothesis.get_timeline()
 
         # reference where short intra-label gaps are removed
         filled = Timeline()
@@ -112,7 +119,7 @@ class SegmentationCoverage(BaseMetric):
 
         return reference_partition, hypothesis_partition
 
-    def _process(self, reference, hypothesis):
+    def _process(self, reference: Annotation, hypothesis: Annotation) -> Details:
 
         detail = self.init_components()
 
@@ -128,14 +135,15 @@ class SegmentationCoverage(BaseMetric):
         return COVERAGE_NAME
 
     @classmethod
-    def metric_components(cls):
+    def metric_components(cls) -> MetricComponents:
         return [PTY_CVG_TOTAL, PTY_CVG_INTER]
 
-    def compute_components(self, reference, hypothesis, **kwargs):
+    def compute_components(self, reference: Annotation,
+                           hypothesis: Union[Annotation, Timeline], **kwargs):
         reference, hypothesis = self._preprocess(reference, hypothesis)
         return self._process(reference, hypothesis)
 
-    def compute_metric(self, detail):
+    def compute_metric(self, detail: Details) -> float:
         return detail[PTY_CVG_INTER] / detail[PTY_CVG_TOTAL]
 
 
@@ -151,10 +159,13 @@ class SegmentationPurity(SegmentationCoverage):
     """
 
     @classmethod
-    def metric_name(cls):
+    def metric_name(cls) -> str:
         return PURITY_NAME
 
-    def compute_components(self, reference, hypothesis, **kwargs):
+    # TODO : Use type from parent class
+    def compute_components(self, reference: Annotation,
+                           hypothesis: Union[Annotation, Timeline],
+                           **kwargs) -> Details:
         reference, hypothesis = self._preprocess(reference, hypothesis)
         return self._process(hypothesis, reference)
 
@@ -186,7 +197,8 @@ class SegmentationPurityCoverageFMeasure(SegmentationCoverage):
         super(SegmentationPurityCoverageFMeasure, self).__init__(tolerance=tolerance, **kwargs)
         self.beta = beta
 
-    def _process(self, reference, hypothesis):
+    def _process(self, reference: Annotation,
+                 hypothesis: Union[Annotation, Timeline]) -> Details:
         reference, hypothesis = self._preprocess(reference, hypothesis)
 
         detail = self.init_components()
@@ -202,32 +214,35 @@ class SegmentationPurityCoverageFMeasure(SegmentationCoverage):
 
         return detail
 
-    def compute_components(self, reference, hypothesis, **kwargs):
+    def compute_components(self, reference: Annotation,
+                           hypothesis: Union[Annotation, Timeline],
+                           **kwargs) -> Details:
         return self._process(reference, hypothesis)
 
-    def compute_metric(self, detail):
+    def compute_metric(self, detail: Details) -> float:
         _, _, value = self.compute_metrics(detail=detail)
         return value
 
-    def compute_metrics(self, detail=None):
+    def compute_metrics(self, detail: Optional[Details] = None) \
+            -> Tuple[float, float, float]:
         detail = self.accumulated_ if detail is None else detail
 
         purity = \
             1. if detail[PTY_TOTAL] == 0. \
-            else detail[PTY_INTER] / detail[PTY_TOTAL]
+                else detail[PTY_INTER] / detail[PTY_TOTAL]
 
         coverage = \
             1. if detail[CVG_TOTAL] == 0. \
-            else detail[CVG_INTER] / detail[CVG_TOTAL]
+                else detail[CVG_INTER] / detail[CVG_TOTAL]
 
         return purity, coverage, f_measure(purity, coverage, beta=self.beta)
 
     @classmethod
-    def metric_name(cls):
+    def metric_name(cls) -> str:
         return PURITY_COVERAGE_NAME
 
     @classmethod
-    def metric_components(cls):
+    def metric_components(cls) -> MetricComponents:
         return [PTY_TOTAL, PTY_INTER, CVG_TOTAL, CVG_INTER]
 
 
@@ -268,10 +283,13 @@ class SegmentationPrecision(UEMSupportMixin, BaseMetric):
 
     def __init__(self, tolerance=0., **kwargs):
 
-        super(SegmentationPrecision, self).__init__(**kwargs)
+        super().__init__(**kwargs)
         self.tolerance = tolerance
 
-    def compute_components(self, reference, hypothesis, **kwargs):
+    def compute_components(self,
+                           reference: Union[Annotation, Timeline],
+                           hypothesis: Union[Annotation, Timeline],
+                           **kwargs) -> Details:
 
         # extract timeline if needed
         if isinstance(reference, Annotation):
@@ -282,7 +300,7 @@ class SegmentationPrecision(UEMSupportMixin, BaseMetric):
         detail = self.init_components()
 
         # number of matches so far...
-        nMatches = 0.  # make sure it is a float (for later ratio)
+        n_matches = 0.  # make sure it is a float (for later ratio)
 
         # number of boundaries in reference and hypothesis
         N = len(reference) - 1
@@ -297,13 +315,13 @@ class SegmentationPrecision(UEMSupportMixin, BaseMetric):
             return detail
 
         # reference and hypothesis boundaries
-        refBoundaries = [segment.end for segment in reference][:-1]
-        hypBoundaries = [segment.end for segment in hypothesis][:-1]
+        ref_boundaries = [segment.end for segment in reference][:-1]
+        hyp_boundaries = [segment.end for segment in hypothesis][:-1]
 
         # temporal delta between all pairs of boundaries
         delta = np.zeros((N, M))
-        for r, refBoundary in enumerate(refBoundaries):
-            for h, hypBoundary in enumerate(hypBoundaries):
+        for r, refBoundary in enumerate(ref_boundaries):
+            for h, hypBoundary in enumerate(hyp_boundaries):
                 delta[r, h] = abs(refBoundary - hypBoundary)
 
         # make sure boundaries too far apart from each other cannot be matched
@@ -317,7 +335,7 @@ class SegmentationPrecision(UEMSupportMixin, BaseMetric):
         # while there are still boundaries to match
         while h < np.inf:
             # increment match count
-            nMatches += 1
+            n_matches += 1
 
             # find boundaries to match
             k = np.argmin(delta)
@@ -331,10 +349,10 @@ class SegmentationPrecision(UEMSupportMixin, BaseMetric):
             # update minimum value in delta
             h = np.amin(delta)
 
-        detail[PR_MATCHES] = nMatches
+        detail[PR_MATCHES] = n_matches
         return detail
 
-    def compute_metric(self, detail):
+    def compute_metric(self, detail: Details) -> float:
 
         numerator = detail[PR_MATCHES]
         denominator = detail[PR_BOUNDARIES]
@@ -379,6 +397,8 @@ class SegmentationRecall(SegmentationPrecision):
     def metric_name(cls):
         return RECALL_NAME
 
-    def compute_components(self, reference, hypothesis, **kwargs):
+    def compute_components(self, reference: Union[Annotation, Timeline],
+                           hypothesis: Union[Annotation, Timeline],
+                           **kwargs) -> Details:
         return super(SegmentationRecall, self).compute_components(
             hypothesis, reference)